backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
pydantic==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
Pillow==10.4.0
python-multipart==0.0.12
